move client menu to roles module
